refactor(post): split out Tags component from Post template
- adjust image size in Post template
